My HW 2
on a scale from 0 to 5, how complete is this homework?
4
on a scale from 0 to 5, how comfortable were you with this assignment?
4
what was a win you had with this assignment?
practice for what we study
what was a challenge you had with this assignment?
the time
other stuff on your mind? work with anyone? tell us about it!