My answer
on a scale from 0 to 5, how complete is this homework? 5
on a scale from 0 to 5, how comfortable were you with this assignment? 5
what was a win you had with this assignment? good at loops and array
what was a challenge you had with this assignment? none
other stuff on your mind? work with anyone? tell us about it! (None)