homework done !
on a scale from 0 to 5, how complete is this homework? 5
on a scale from 0 to 5, how comfortable were you with this assignment? 4
what was a win you had with this assignment? nothing
what was a challenge you had with this assignment? write code!
other stuff on your mind? work with anyone? tell us about it!  thank you