homework Done
on a scale from 0 to 5, how complete is this homework? 5
on a scale from 0 to 5, how comfortable were you with this assignment? 4
what was a win you had with this assignment? all of it
what was a challenge you had with this assignment? random variables.
other stuff on your mind? work with anyone? tell us about it! ask friends about some points and help each other